Added details in summary
The field "details" contains very useful information to take decision in case of conflicts.
While in the question section it was possible to show the details , in the summary this option was missing .
With this change we can immediately see the details and avoid mistake in the merge phase. 

The ID has been remove (only from the UI layer) because it was not relevant in the decision process and it took up a lot of UI space.